TASK: Ensure file path is returned before checking

+++ Classes/Resource/Handler/ImageBuilderResource.php
@@ -97,7 +97,7 @@ class ImageBuilderResource implements RemoteResourceInterface
         $gifBuilder = GeneralUtility::makeInstance(GifBuilder::class);
         $gifBuilder->start($fileArray, []);
         $theImage = $gifBuilder->gifBuild()?->getFullPath();
-        if (file_exists($theImage)) {
+        if ($theimage && file_exists($theImage)) {
             $content = file_get_contents($theImage);
             unlink($theImage);
         }
